Moved WeekFragment under weektimes package
